Fix changed id from int to Long

+++ src/main/java/nl/fontys/realestateproject/business/TransactionService.java
@@ -9,7 +9,7 @@ public interface TransactionService {
 
     GetAllTransactionResponse getAllTransactions();
 
-    GetAllTransactionResponse getTransactionsByCustomerId(int customerId);
+    GetAllTransactionResponse getTransactionsByCustomerId(Long customerId);
 
-    GetAllTransactionResponse getTransactionsByPropertyId(int propertyId);
+    GetAllTransactionResponse getTransactionsByPropertyId(Long propertyId);
 }
